Make tests run on node v4+
node 4+ provides os.homedir() implementation. 'os-homedir' module favours it
over own implementation. This module is dependency of 'osenv' that we
use for env data resolution.

We still need to have HOME in our 'fake' env. 'appdirectory' expects it.
(at least the version we're using does - no access to internet to check
if there is a later version that doesn't)