Remove NVM version dependency check
If NVM is installed from git or homebrew then the package.json for NVM
may not exist, resulting in bosco failing to work.  The version check
was to ensure at least nvm 0.21.0 was installed for some required API changes.
0.21.0 was released over 3 years ago so it's safe to assume new users
will definitely have a newer version and those who don't will be able to
figure it out.