Tinkering with bosco
This is where we got to in the 'mob on bosco' session during the
open space. It's totally broken cos we're mid-refactor of swapping
boscoRepo.assets to realAssets.

We had hard-coded the assets to a single js file, with the aim to
later get the files from the dist folder (and eventually any
folder configured in the bosco json file).

This requires the following in the bosco-service.json of the
app/service you're building:
  "generateAssetsFromDist": true

Then you can try the changes in this commit by running (from the
folder where you have that app/service, and assuming both that and
bosco and checked out in the same dir):
  node ../bosco/bin/bosco.js cdn